+++ src/barcode-scanner-service/Dockerfile.python
