+++ src/barcode-scanner-service/Makefile
