+++ src/barcode-scanner-service/docker-compose.yml
