+++ src/barcode-scanner-service/requirements.txt
